pkg/migration tidy: lint runner methods -> runner.go; DSN helpers -> dbconn
Two small organizational cleanups; no behavior change to the migration path.

lint.go -> runner.go
  Move the (r *Runner) lint and (r *Runner) getCreateTable methods from lint.go
  into runner.go so all Runner methods live in one file. lint.go now holds only
  the package-level lint helpers (defaultLinterSettings, printLinters).

DSN helpers -> pkg/dbconn (and unify password masking)
  Move splitReplicaDSNs and maskPasswordInDSN out of pkg/migration/utils.go into
  pkg/dbconn as SplitDSNs and RedactDSN, and delete the now-empty utils.go.
  datasync had its own redactDSN with different (less informative) behavior; both
  migration and datasync now call the single dbconn.RedactDSN. The unified
  RedactDSN keeps migration's robust ParseDSN-based password-only masking (so the
  user/host stay visible in logs) and adds datasync's safe "redact everything
  before @" fallback for DSNs the driver can't parse, so a password is never
  logged verbatim. Tests moved to pkg/dbconn/dsn_test.go.
