+++ pkg/migration/runner.go
@@ -6,6 +6,7 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
+	"strings"
 	"sync"
 	"time"
 
@@ -17,8 +18,10 @@ import (
 	"github.com/block/spirit/pkg/checksum"
 	"github.com/block/spirit/pkg/copier"
 	"github.com/block/spirit/pkg/dbconn"
+	"github.com/block/spirit/pkg/lint"
 	"github.com/block/spirit/pkg/metrics"
 	"github.com/block/spirit/pkg/migration/check"
+	"github.com/block/spirit/pkg/statement"
 	"github.com/block/spirit/pkg/status"
 	"github.com/block/spirit/pkg/table"
 	"github.com/block/spirit/pkg/throttler"
@@ -238,7 +241,7 @@ func (r *Runner) Run(ctx context.Context) error {
 	r.dbConfig.MaxOpenConnections = r.migration.Threads + r.migration.WriteThreads + r.controlPlaneConns()
 	r.db, err = dbconn.New(r.dsn(), r.dbConfig)
 	if err != nil {
-		return fmt.Errorf("failed to connect to main database (DSN: %s): %w", maskPasswordInDSN(r.dsn()), err)
+		return fmt.Errorf("failed to connect to main database (DSN: %s): %w", dbconn.RedactDSN(r.dsn()), err)
 	}
 
 	// Run linting if --lint or --lint-only is specified.
@@ -513,6 +516,78 @@ func (r *Runner) runChecks(ctx context.Context, scope check.ScopeFlag) error {
 		}
 	}
 	return nil
+}
+
+func (r *Runner) lint(ctx context.Context) error {
+	var createTables []*statement.CreateTable
+	var alterTables []*statement.AbstractStatement
+	config := lint.Config{
+		Enabled:  make(map[string]bool),
+		Settings: defaultLinterSettings,
+	}
+
+	if err := printLinters(config); err != nil {
+		return err
+	}
+
+	for _, change := range r.changes {
+		// Collect ALTER TABLE statements and the CREATE TABLEs for the tables they reference
+		if change.stmt.IsAlterTable() {
+			alterTables = append(alterTables, change.stmt)
+
+			ct, err := r.getCreateTable(ctx, change.stmt.Schema, change.stmt.Table)
+			if err != nil {
+				return err
+			}
+			createTables = append(createTables, ct)
+		}
+
+		// If the migration creates a table, we need to collect that CREATE TABLE as well
+		if change.stmt.IsCreateTable() {
+			ct, err := change.stmt.ParseCreateTable()
+			if err != nil {
+				return err
+			}
+			createTables = append(createTables, ct)
+		}
+	}
+
+	var errs []error
+
+	violations, err := lint.RunLinters(createTables, alterTables, config)
+	if err != nil {
+		errs = append(errs, err)
+	}
+
+	for _, v := range violations {
+		if v.Severity == lint.SeverityError {
+			errs = append(errs, errors.New(v.String()))
+		}
+		fmt.Println(v)
+	}
+
+	if len(errs) > 0 {
+		return errors.Join(errs...)
+	}
+	return nil
+}
+
+func (r *Runner) getCreateTable(ctx context.Context, db string, tbl string) (*statement.CreateTable, error) {
+	// Escape backticks in db and tbl names to be extra pedantic
+	db = strings.ReplaceAll(db, "`", "``")
+	tbl = strings.ReplaceAll(tbl, "`", "``")
+	sql := fmt.Sprintf("show create table `%s`.`%s`", db, tbl)
+
+	row := r.db.QueryRowContext(ctx, sql)
+	var createTable string
+	if err := row.Scan(&tbl, &createTable); err != nil {
+		return nil, err
+	}
+	stmt, err := statement.ParseCreateTable(createTable)
+	if err != nil {
+		return nil, err
+	}
+	return stmt, nil
 }
 
 func (r *Runner) dsn() string {
@@ -820,7 +895,7 @@ func (r *Runner) setupThrottler(ctx context.Context) error {
 // throttler per replica. Replica connections are tracked on the runner so
 // they get closed alongside the main DB.
 func (r *Runner) buildReplicaThrottlers() ([]throttler.Throttler, error) {
-	dsns := splitReplicaDSNs(r.migration.ReplicaDSN)
+	dsns := dbconn.SplitDSNs(r.migration.ReplicaDSN)
 	if len(dsns) == 0 {
 		return nil, fmt.Errorf("--replica-dsn was specified but contains no valid DSNs: %q", r.migration.ReplicaDSN)
 	}
@@ -842,7 +917,7 @@ func (r *Runner) buildReplicaThrottlers() ([]throttler.Throttler, error) {
 		enhancedDSN, err := dbconn.EnhanceDSNWithTLS(dsn, replicaDBConfig)
 		if err != nil {
 			r.logger.Warn("could not enhance replica DSN with TLS settings",
-				"dsn", maskPasswordInDSN(dsn),
+				"dsn", dbconn.RedactDSN(dsn),
 				"error", err,
 			)
 			enhancedDSN = dsn
@@ -851,14 +926,14 @@ func (r *Runner) buildReplicaThrottlers() ([]throttler.Throttler, error) {
 		replicaDB, err := dbconn.NewWithConnectionType(enhancedDSN, replicaDBConfig, "replica database")
 		if err != nil {
 			_ = r.closeReplicas()
-			return nil, fmt.Errorf("failed to connect to replica database (DSN: %s): %w", maskPasswordInDSN(dsn), err)
+			return nil, fmt.Errorf("failed to connect to replica database (DSN: %s): %w", dbconn.RedactDSN(dsn), err)
 		}
 		r.replicas = append(r.replicas, replicaDB)
 
 		replicaThrottler, err := throttler.NewReplicationThrottler(replicaDB, r.migration.ReplicaMaxLag, r.logger)
 		if err != nil {
 			_ = r.closeReplicas()
-			return nil, fmt.Errorf("could not create replication throttler (DSN: %s): %w", maskPasswordInDSN(dsn), err)
+			return nil, fmt.Errorf("could not create replication throttler (DSN: %s): %w", dbconn.RedactDSN(dsn), err)
 		}
 		throttlers = append(throttlers, replicaThrottler)
 	}
